all endpoints are functional

+++ src/main/java/com/tyss/dashboard/branch/mng/services/BranchServicesImpl.java
@@ -1,6 +1,7 @@
 package com.tyss.dashboard.branch.mng.services;
 
 import java.util.List;
+import java.util.UUID;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.mongodb.core.MongoTemplate;
@@ -11,9 +12,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import com.tyss.dashboard.branch.mng.entities.BranchEntity;
-import com.tyss.dashboard.branch.mng.model.BranchDto;
 import com.tyss.dashboard.branch.mng.repositories.BranchRepository;
-
 
 @Service
 public class BranchServicesImpl implements BranchServices {
@@ -25,56 +24,49 @@ public class BranchServicesImpl implements BranchServices {
 	MongoTemplate mongoTemplate;
 
 	@Override
-	public ResponseEntity<String> addBranch(BranchEntity branchEntity) {
+	public ResponseEntity<BranchEntity> addBranch(BranchEntity branchEntity) {
 
 		if (mongoTemplate
 				.find(new Query().addCriteria(Criteria.where("branchName").is(branchEntity.getBranchName()).and("city")
 						.is(branchEntity.getCity()).and("type").is(branchEntity.getType())), BranchEntity.class)
 				.size() == 0) {
-			branchEntity.setId(branchEntity.hashCode());
-			branchEntity.setBranchName(branchEntity.getBranchName().toLowerCase());
+			branchEntity.setId(UUID.randomUUID().toString());
+			branchEntity.setBranchName(branchEntity.getBranchName());
 			branchEntity.setBranchManager(branchEntity.getBranchManager().toLowerCase());
 			branchRepository.save(branchEntity);
 
-			return ResponseEntity.status(HttpStatus.CREATED).body("Branch added successfully!!");
+			return ResponseEntity.status(HttpStatus.OK).body(branchEntity);
 
 		}
 
-		return ResponseEntity.status(HttpStatus.CONFLICT).body("Branch already exists!!");
+		return ResponseEntity.status(HttpStatus.CONFLICT).body(null);
 	}
 
 	@Override
-	public ResponseEntity<String> deleteBranch(BranchDto branchDto) {
+	public ResponseEntity<String> deleteBranch(String branchID) {
+
 		BranchEntity branchEntity = null;
-		try {
-			branchEntity = mongoTemplate
-					.find(new Query().addCriteria(Criteria.where("branchName").is(branchDto.getBranchName()).and("city")
-							.is(branchDto.getCity()).and("type").is(branchDto.getType())), BranchEntity.class)
-					.get(0);
+
+		branchEntity = branchRepository.findById(branchID).get();
+		if (branchEntity != null) {
 
 			branchRepository.delete(branchEntity);
-
 			return ResponseEntity.status(HttpStatus.ACCEPTED).body("Branch deleted!!");
-		} catch (IndexOutOfBoundsException ie) {
-			return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Branch not found!!");
-
 		}
+
+		return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Branch not found!!");
+
 	}
 
 	@Override
-	public ResponseEntity<BranchEntity> viewBranch(String branchName, String city, String type) {
-		BranchEntity branchEntity = null;
-		try {
-			branchEntity = mongoTemplate.find(
-					new Query().addCriteria(
-							Criteria.where("branchName").is(branchName).and("city").is(city).and("type").is(type)),
-					BranchEntity.class).get(0);
-
-			return ResponseEntity.status(HttpStatus.FOUND).body(branchEntity);
-		} catch (IndexOutOfBoundsException ie) {
-			return ResponseEntity.status(HttpStatus.NOT_FOUND).body(branchEntity);
-
+	public ResponseEntity<BranchEntity> viewBranch(String branchName) {
+		
+		BranchEntity branchEntity = branchRepository.findByBranchName(branchName);
+		if (branchEntity != null) {
+			return ResponseEntity.status(HttpStatus.OK).body(branchEntity);
 		}
+		return ResponseEntity.status(HttpStatus.NOT_FOUND).body(branchEntity);
+
 	}
 
 	@Override
@@ -88,7 +80,5 @@ public class BranchServicesImpl implements BranchServices {
 
 		return ResponseEntity.status(HttpStatus.OK).body(branchRepository.findAllByCity(city));
 	}
-
-	
 
 }
